[frontend] chore: Реорганизация файловой структуры (#2)
