+++ frontend/src/Screens.tsx
@@ -1,23 +1,23 @@
 import { View, StyleSheet, useWindowDimensions, Text } from "react-native";
-import { ScreenType } from "../../App";
+import { ScreenType } from "../App";
 import Animated, {
   useAnimatedStyle,
   useSharedValue,
   withTiming,
 } from "react-native-reanimated";
-import { news, NewsScreen } from "../components/NewsScreen";
-import { Encyclopedia } from "../components/Encyclopedia";
-import { Maps } from "../components/Maps";
-import { Search } from "../components/Search";
+import { news, NewsScreen } from "./screens/NewsScreen";
+import { Encyclopedia } from "./screens/Encyclopedia";
+import { Maps } from "./screens/Maps";
+import { Search } from "./screens/Search";
 
 type Props = {
   currentScreen: ScreenType;
 };
 
 const screens: { [key in ScreenType]: JSX.Element } = {
-  Карта: <Maps/>,
+  Карта: <Maps />,
   Энциклопедия: <Encyclopedia />,
-  Поиск: <Search/>,
+  Поиск: <Search />,
   Новости: <NewsScreen data={news} />,
   Профиль: <Text>Я profile</Text>,
 };
@@ -26,11 +26,15 @@ export const Screens: React.FC<Props> = (props) => {
   const { width } = useWindowDimensions();
   const Screen: React.FC<React.PropsWithChildren> = ({ children }) => (
     <View style={[styles.screen, { width }]}>{children}</View>
-  ); 
+  );
 
   return (
     <View style={styles.wrapper}>
-        {<Screen key={props.currentScreen}>{screens[props.currentScreen as ScreenType]}</Screen>}
+      {
+        <Screen key={props.currentScreen}>
+          {screens[props.currentScreen as ScreenType]}
+        </Screen>
+      }
     </View>
   );
 };
