+++ frontend/src/screens/NewsScreen.tsx
